+++ R/sim_dist.R
@@ -156,17 +156,26 @@ sim_ays_covar <- function(sd = 2.8, range = 300, lambda = 1, model = "matern",
 
 #' Define relationships with covariates
 #'
-#' @description  Closure to be used in \code{\link{sim_distribution}}
+#' @description  Closure to be used in \code{\link{sim_distribution}}. Form is based on the bi-gaussian function described here: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2993707/.
 #'
 #' @param alpha,mu,sigma  Parameters that control the shape of the parabola. Can be one value or
 #'                        a vector of equal length to the number of ages in the simulation
 #'                        (e.g. age-specific depth associations can be specified).
+#' @param sigma_right     Optional parameter to impose asymmetry by supplying a sigma parameter for the right side.
+#'                        If used, `sigma` will be used to define the width of the left side. Ignored if `NULL`.
+#' @param log_space       Should shape of the parabola be defined in log space? If `TRUE`, logged parameters are
+#'                        assumed to be supplied and x values used in the parabola equation are log transformed.
+#'                        This allows a more lognormal curve to be defined and, hence, allows a heavier tail and it
+#'                        forces very low values near zero.
 #' @param plot            Produce a simple plot of the simulated values?
 #'
 #' @examples
 #'
-#' parabola_fun <- sim_parabola(alpha = 25, mu = 50, sigma = 5, plot = TRUE)
+#' parabola_fun <- sim_parabola(mu = 50, sigma = 5, plot = TRUE)
 #' parabola_fun(x = 0:100)
+#'
+#' parabola_fun <- sim_parabola(mu = log(40), sigma = 0.5, log_space = TRUE, plot = TRUE)
+#' parabola_fun(x = 1:1000)
 #'
 #' parabola_fun <- sim_parabola(mu = c(50, 120), sigma = c(5, 3), plot = TRUE)
 #' parabola_fun(x = rep(1:200, 2), age = rep(c(1, 2), each = 200))
@@ -174,7 +183,8 @@ sim_ays_covar <- function(sd = 2.8, range = 300, lambda = 1, model = "matern",
 #' @rdname sim_parabola
 #' @export
 
-sim_parabola <- function(alpha = 0, mu = 200, sigma = 70, plot = FALSE) {
+sim_parabola <- function(alpha = 0, mu = 200, sigma = 70, sigma_right = NULL,
+                         log_space = FALSE, plot = FALSE) {
 
   function(x = NULL, age = NULL) {
 
@@ -195,10 +205,17 @@ sim_parabola <- function(alpha = 0, mu = 200, sigma = 70, plot = FALSE) {
       i <- rep(1, length(x))
     }
 
-    y <- alpha[i] - (((x - mu[i])^2) / (2 * sigma[i] ^ 2))
+    if (is.null(sigma_right)) sigma_right <- sigma
+    if (log_space) mx <- log(x) else mx <- x
+
+    y <- ifelse(mx < mu[i], alpha[i] - (((mx - mu[i])^2) / (2 * sigma[i] ^ 2)),
+                            alpha[i] - (((mx - mu[i])^2) / (2 * sigma_right[i] ^ 2)))
+
     if (plot) {
-      plot(x, exp(y), main = "sim_parabola", col = i)
+      plot(x, y, main = "sim_parabola", col = i)
     }
+
+    plot_ly(x = x, y = exp(y)) %>% add_lines()
     y
 
   }