+++ NEWS.md
@@ -9,4 +9,4 @@
 # SimSurvey 0.1.2.009
 
 * Simplify make_grid to minimize conflicts between strat splits and divisions
-* Allow a vector of age-specific parameters to be supplied to sim_parabola
+* Allow a vector of age-specific parameters to be supplied to sim_parabola plus add some options for defining a more asymetric parabola